# V2 Beta A
## Removed
### Volume keys to configure
This was a pain, for some it doesn't work

## Added
### PreConfiged File
Each release file will have it's own config

### More Configs
12 Different Combinations Added. You can add mode under config/

### Fail Safe
No Idea on how well it'll work, but if the service file detects an incomplete start up of the last swapon, it will permenently disable itself.
To re-enable, it you have to re-install the module.

## Log File
is persistant now, but if you uninstall the module, it will get deleted